Behebe kritische Compiler-Fehler und Null-Safety Issues
- CocaPlantBlock: Default-Case für CocaType Switch (PERUANISCH)
- DialogueHelper: Korrekte Methodensignaturen für startDialogue/endDialogue
- NPCInteractionManager: Wallet-Transfer mit korrektem int-Typ
- NPCLifeSystemIntegration: getNpcName() statt getName()

+++ src/main/java/de/rolandsw/schedulemc/npc/life/social/NPCInteractionManager.java
@@ -191,7 +191,7 @@ public class NPCInteractionManager {
         }
 
         // Wallet-Transfer
-        long buyerWallet = buyer.getNpcData().getWallet();
+        int buyerWallet = buyer.getNpcData().getWallet();
         if (buyerWallet < amount) {
             return false; // Nicht genug Geld
         }
